pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-zero"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "anthropic>=0.47.1",
  "docker>=7.1.0",
  "google-genai>=1.4.0",
  "openai>=1.0.0",
  "python-dotenv>=0.21.0",
]
license = { text = "MIT" }
keywords = ["agents", "multi-agent", "orchestration", "AI"]

[tool.setuptools]
packages = ["agents_manager"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
